requirements.txt
redis
hiredis
rdflib
Agora-Service-Provider

